+++ CHANGELOG.md
@@ -6,7 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - All() queries with attribute selection would throw an error
-- fixed entityset batching when adding arrays of entities/components 
+- fixed entityset batching when adding arrays of entities/components
+
+### Changed
+- removed babel from package.json - this appears to confuse react-native 
 
 ## [4.0.6] - 2016-11-20